+++ utils/compare.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 #include <stack>
 #include <string>
-#include "../tree/node.h"
+
 using namespace std;
 
 int get_common_prefix_len(const char *a, const char *b, int alen, int blen) {
@@ -48,32 +48,3 @@ int char_cmp_new(const char *a, const char *b, int alen, int blen) { // gaurunte
     #endif
 }
 
-#ifdef PV
-long word_cmp(Stdhead* header,const char* key, int keylen) {
-    // char word[8] = {0};
-    // char prefix[8] = {0};
-    // for (int i = 0; i < PV_SIZE; i++) 
-    //     prefix[i] = header->key_prefix[PV_SIZE - 1 - i];
-    // for (int i = 0; i < min(keylen, PV_SIZE); i++)
-    //     word[i] = key[min(keylen, PV_SIZE) - 1 - i];
-    // return *(long*)word - *(long*)prefix;
-    int cmp_len = min(PV_SIZE, keylen);
-    // int idx = *matchp;
-    for (int idx = 0; idx < cmp_len; ++idx) {
-        int cmp = key[idx] - header->key_prefix[idx];
-        if (cmp != 0)
-            return cmp;
-    }
-    /* Contents are equal up to the smallest length. */
-    return 0;
-}
-
-inline long pvComp(Stdhead* header,const char* key, int keylen, Node *cursor) {
-    long cmp = word_cmp(header, key, keylen);
-    if (cmp == 0) {
-        cmp = char_cmp_new(key, PageOffset(cursor, header->key_offset),
-                            keylen, header->key_len);
-    }
-    return cmp;
-}
-#endif
